tests/aockt/y2023/Y2023D10Test.kt
package aockt.y2023

import io.github.jadarma.aockt.test.AdventDay
import io.github.jadarma.aockt.test.AdventSpec

@AdventDay(2023, 10, "Pipe Maze")
class Y2023D10Test : AdventSpec<Y2023D10>({
    val officialTestcase1 = """
        .....
        .S-7.
        .|.|.
        .L-J.
        .....
    """.trimIndent()
    val officialTestcase2 = """
        ..F7.
        .FJ|.
        SJ.L7
        |F--J
        LJ...
    """.trimIndent()

    partOne {
        officialTestcase1 shouldOutput 4
        officialTestcase2 shouldOutput 8
    }

    partTwo {
        """
            ...........
            .S-------7.
            .|F-----7|.
            .||.....||.
            .||.....||.
            .|L-7.F-J|.
            .|..|.|..|.
            .L--J.L--J.
            ...........
        """.trimIndent() shouldOutput 4
        """
            ..........
            .S------7.
            .|F----7|.
            .||....||.
            .||....||.
            .|L-7F-J|.
            .|..||..|.
            .L--JL--J.
            ..........
        """.trimIndent() shouldOutput 4
        """
            .F----7F7F7F7F-7....
            .|F--7||||||||FJ....
            .||.FJ||||||||L7....
            FJL7L7LJLJ||LJ.L-7..
            L--J.L7...LJS7F-7L7.
            ....F-J..F7FJ|L7L7L7
            ....L7.F7||L7|.L7L7|
            .....|FJLJ|FJ|F7|.LJ
            ....FJL-7.||.||||...
            ....L---J.LJ.LJLJ...
        """.trimIndent() shouldOutput 8
    }

})
